+++ cmd/gen/circleci/runner.go
@@ -89,6 +89,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, _ []string) error 
 			Language:                r.flag.Language,
 			Flavours:                r.flag.Flavours,
 			SkipATS:                 r.flag.SkipATS,
+			ATSBranchOnly:           r.flag.ATSBranchOnly,
 			HasDockerfile:           hasDockerfile,
 			AppCatalog:              r.flag.AppCatalog,
 			AppCatalogTest:          r.flag.AppCatalogTest,